Fix Typescript/Closure errors
